Reworked approach to handling templateEditor state
This commit gets rid of the templateEditorStore and replaces it with a composable serving the same purpose.
The reason for doing this is that it proved difficult to handle the template in the editor as separate from the other templates and still keep the data up to date, as well as editing an existing template instead of creating a new one.